requirements.txt
django==2.2
mysqlclient
pillow
requests
django-simple-captcha